requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.0  # Legacy driver (has encoding issues on Windows)
psycopg[binary]>=3.1.0  # Modern driver (psycopg3) - better Windows support
psycopg-pool>=3.2.0  # Connection pooling for psycopg3
sqlalchemy>=2.0.0
duckdb>=0.9.0
redis>=5.0.0

# Message Queue
kafka-python>=2.0.2
# Alternative: redpanda-python (if using Redpanda)

# Data Processing
polars>=0.19.0
pandas>=2.0.0

# ETL & Transformation
dbt-core>=1.6.0
dbt-postgres>=1.6.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

